Add handlers for Chain of Responsibility pattern in generative extraction
